Move verify_config to config.py
