+++ src/purewebm/config.py
@@ -1,0 +1,22 @@
+# Copyright (c) 2022 4ndrs <andres.degozaru@gmail.com>
+# SPDX-License-Identifier: MIT
+"""Module for the handling of the configuration folder"""
+
+import os
+import sys
+
+from . import CONFIG_PATH
+
+
+def verify_config():
+    """Checks the configuration folder, creates it if it doesn't exist"""
+    if not CONFIG_PATH.exists():
+        try:
+            CONFIG_PATH.mkdir(parents=True)
+        except PermissionError:
+            print(
+                "Unable to create the configuration folder "
+                f"{CONFIG_PATH}, permission denied",
+                file=sys.stderr,
+            )
+            sys.exit(os.EX_CANTCREAT)
